refactor datastore package naming and consistency
