+++ elide-datastore/elide-datastore-multiplex/src/main/java/com/yahoo/elide/datastores/multiplex/MultiplexTransaction.java
@@ -1,0 +1,131 @@
+/*
+ * Copyright 2015, Yahoo Inc.
+ * Licensed under the Apache License, Version 2.0
+ * See LICENSE file in project root for terms.
+ */
+package com.yahoo.elide.datastores.multiplex;
+
+import com.yahoo.elide.core.DataStore;
+import com.yahoo.elide.core.DataStoreTransaction;
+import com.yahoo.elide.core.FilterScope;
+import com.yahoo.elide.core.exceptions.InvalidCollectionException;
+import com.yahoo.elide.core.filter.Predicate;
+import com.yahoo.elide.security.User;
+
+import java.io.IOException;
+import java.io.Serializable;
+import java.util.Collection;
+import java.util.LinkedHashMap;
+import java.util.Set;
+
+/**
+ * Multiplex transaction handler.  Process each sub-database transactions within a single transaction.
+ * If any commit fails in process, reverse any commits already completed.
+ */
+public abstract class MultiplexTransaction implements DataStoreTransaction {
+    protected final LinkedHashMap<DataStore, DataStoreTransaction> transactions;
+    protected final MultiplexManager multiplexManager;
+    protected final DataStoreTransaction lastDataStoreTransaction;
+
+    /**
+     * Multiplex transaction handler.
+     * @param multiplexManager associated manager
+     */
+    public MultiplexTransaction(MultiplexManager multiplexManager) {
+        this.multiplexManager = multiplexManager;
+        this.transactions = new LinkedHashMap<>(multiplexManager.dataStores.size());
+
+        // create each subordinate transaction
+        DataStoreTransaction transaction = null;
+        for (DataStore dataStore : multiplexManager.dataStores) {
+            transaction = beginTransaction(dataStore);
+            transactions.put(dataStore, transaction);
+        }
+        lastDataStoreTransaction = transaction;
+    }
+
+    protected abstract DataStoreTransaction beginTransaction(DataStore dataStore);
+
+    @Override
+    public User accessUser(Object opaqueUser) {
+        User user = new User(opaqueUser);
+        for (DataStore dataStore : multiplexManager.dataStores) {
+            DataStoreTransaction transaction = transactions.get(dataStore);
+            user = transaction.accessUser(user.getOpaqueUser());
+        }
+        return user;
+    }
+
+    @Override
+    public <T> T createObject(Class<T> createObject) {
+        return getTransaction(createObject).createObject(createObject);
+    }
+
+    @Override
+    public <T> T loadObject(Class<T> loadClass, Serializable id) {
+        return getTransaction(loadClass).loadObject(loadClass, id);
+    }
+
+    @Override
+    public <T> Iterable<T> loadObjects(Class<T> loadClass) {
+        return getTransaction(loadClass).loadObjects(loadClass);
+    }
+
+    @Override
+    public <T> Iterable<T> loadObjects(Class<T> entityClass, FilterScope<T> filterScope) {
+        return getTransaction(entityClass).loadObjects(entityClass, filterScope);
+    }
+
+    @Override
+    public <T> Collection filterCollection(Collection collection, Class<T> entityClass, Set<Predicate> predicates) {
+        return getTransaction(entityClass).filterCollection(collection, entityClass, predicates);
+    }
+
+    @Override
+    public void flush() {
+        transactions.values().forEach(DataStoreTransaction::flush);
+    }
+
+    @Override
+    public void commit() {
+        // flush all before commit
+        flush();
+        transactions.values().forEach(DataStoreTransaction::commit);
+        transactions.clear();
+    }
+
+    @Override
+    public void close() throws IOException {
+
+        IOException cause = null;
+        for (DataStoreTransaction transaction : transactions.values()) {
+            try {
+                transaction.close();
+            } catch (IOException | Error | RuntimeException e) {
+                if (cause != null) {
+                    cause.addSuppressed(e);
+                } else if (e instanceof IOException) {
+                    cause = (IOException) e;
+                } else {
+                    cause = new IOException(e);
+                }
+            }
+        }
+        if (cause != null) {
+            throw cause;
+        }
+    }
+
+    protected DataStoreTransaction getTransaction(Object object) {
+        return getTransaction(object.getClass());
+    }
+
+    protected DataStoreTransaction getTransaction(Class<?> cls) {
+        DataStoreTransaction transaction = transactions.get(this.multiplexManager.getSubManager(cls));
+        if (transaction == null) {
+            Class entityClass = this.multiplexManager.getDictionary().lookupEntityClass(cls);
+            throw new InvalidCollectionException(entityClass == null ? cls.getName() : entityClass.getName());
+        }
+        return transaction;
+    }
+}
