+++ elide-integration-tests/src/test/groovy/com/yahoo/elide/tests/ShareableIT.groovy
@@ -7,7 +7,7 @@ package com.yahoo.elide.tests
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.yahoo.elide.core.HttpStatus
-import com.yahoo.elide.datastore.AHibernateTest
+import com.yahoo.elide.datastores.AHibernateTest
 import org.testng.Assert
 import org.testng.annotations.Test
