+++ elide-integration-tests/src/test/java/com/yahoo/elide/endpoints/TestApplicationResourceConfig.java
@@ -8,7 +8,7 @@ package com.yahoo.elide.endpoints;
 import com.yahoo.elide.Elide;
 import com.yahoo.elide.audit.TestLogger;
 import com.yahoo.elide.core.EntityDictionary;
-import com.yahoo.elide.datastore.AHibernateTest;
+import com.yahoo.elide.datastores.AHibernateTest;
 import com.yahoo.elide.resources.JsonApiEndpoint;
 
 import org.glassfish.hk2.api.Factory;
