+++ elide-integration-tests/src/test/java/example/Filtered.java
@@ -12,7 +12,7 @@ import com.yahoo.elide.annotation.ReadPermission;
 import com.yahoo.elide.annotation.UpdatePermission;
 import com.yahoo.elide.core.PersistentResource;
 import com.yahoo.elide.core.RequestScope;
-import com.yahoo.elide.security.CriteriaCheck;
+import com.yahoo.elide.datastores.hibernate3.security.CriteriaCheck;
 import com.yahoo.elide.security.Role;
 
 import example.Filtered.FilterCheck;
