+++ elide-integration-tests/testngIT.xml
@@ -13,9 +13,9 @@
             <class name="com.yahoo.elide.endpoints.ResourceIT" />
         </classes>
     </test> <!-- Test -->
-    <test name="com.yahoo.elide.datastore.ExampleIT">
+    <test name="com.yahoo.elide.datastores.ExampleIT">
         <classes>
-            <class name="com.yahoo.elide.datastore.ExampleIT" />
+            <class name="com.yahoo.elide.datastores.ExampleIT" />
         </classes>
     </test> <!-- Test -->
     <test name="com.yahoo.elide.endpoints.EmbeddedIT">
